pyproject.toml
[tool.poetry]
name = "colivingcash"
version = "0.1.0"
description = "App for count your money"
authors = ["devkult <devkultanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.4.2"
dishka = "^1.2.0"
faker = "^26.0.0"
fastapi = "^0.111.0"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
aiokafka = "^0.11.0"
aiojobs = "^1.3.0"
pytest-asyncio = "^0.24.0"


[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
